refactor(BanqueSaidi) fix id type in war   -- MS

+++ BanqueSaidi-war/src/java/beans/CompteBancaireMBean.java
@@ -31,12 +31,12 @@ public class CompteBancaireMBean implements Serializable {
     private GestionnaireCompteBancaire gc;
     private CompteBancaire compteBancaire;
     private List<CompteBancaire> listCompteBancaire;
-    private int idCompteACrediter;
+    private Long idCompteACrediter;
     private double montantACrediter;
-    private int idCompteADebiter;
+    private Long idCompteADebiter;
     private double montantADebiter;
-    private int id1;
-    private int id2;
+    private Long id1;
+    private Long id2;
     private double montantTransfert;
     private String message;
     private LazyDataModel<CompteBancaire> modele;
@@ -102,7 +102,7 @@ public class CompteBancaireMBean implements Serializable {
      *
      * @return the value of id2
      */
-    public int getId2() {
+    public Long getId2() {
         return id2;
     }
 
@@ -111,7 +111,7 @@ public class CompteBancaireMBean implements Serializable {
      *
      * @param id2 new value of id2
      */
-    public void setId2(int id2) {
+    public void setId2(Long id2) {
         this.id2 = id2;
     }
 
@@ -138,7 +138,7 @@ public class CompteBancaireMBean implements Serializable {
      *
      * @return the value of id1
      */
-    public int getId1() {
+    public Long getId1() {
         return id1;
     }
 
@@ -147,7 +147,7 @@ public class CompteBancaireMBean implements Serializable {
      *
      * @param id1 new value of id1
      */
-    public void setId1(int id1) {
+    public void setId1(Long id1) {
         this.id1 = id1;
     }
 
@@ -174,7 +174,7 @@ public class CompteBancaireMBean implements Serializable {
      *
      * @return the value of idCompteADebiter
      */
-    public int getIdCompteADebiter() {
+    public Long getIdCompteADebiter() {
         return idCompteADebiter;
     }
 
@@ -183,7 +183,7 @@ public class CompteBancaireMBean implements Serializable {
      *
      * @param idCompteADebiter new value of idCompteADebiter
      */
-    public void setIdCompteADebiter(int idCompteADebiter) {
+    public void setIdCompteADebiter(Long idCompteADebiter) {
         this.idCompteADebiter = idCompteADebiter;
     }
 
@@ -210,7 +210,7 @@ public class CompteBancaireMBean implements Serializable {
      *
      * @return the value of idCompteACrediter
      */
-    public int getIdCompteACrediter() {
+    public Long getIdCompteACrediter() {
         return idCompteACrediter;
     }
 
@@ -219,7 +219,7 @@ public class CompteBancaireMBean implements Serializable {
      *
      * @param idCompteACrediter new value of idCompteACrediter
      */
-    public void setIdCompteACrediter(int idCompteACrediter) {
+    public void setIdCompteACrediter(Long idCompteACrediter) {
         this.idCompteACrediter = idCompteACrediter;
     }
